requirements.txt
playwright==1.30.0
pytest==7.1.2
python-dotenv==0.21.0